requirements.txt
django
google-genai
mysqlclient
pytz
python-dotenv
gunicorn
whitenoise